починила сохранение файлов странслированных
но это не точно
